requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
pygame-ce==2.5.6
PyTMX==3.32
requests==2.32.5
urllib3==2.6.3
